API-2218 code review changes
