+++ app/uk/gov/hmrc/apigateway/connector/impl/ApiDefinitionConnector.scala
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2016 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.apigateway.connector.impl
+
+import javax.inject.{Inject, Singleton}
+
+import play.api.libs.ws.WSClient
+import uk.gov.hmrc.apigateway.connector.ServiceConnector
+import uk.gov.hmrc.apigateway.exception.GatewayError.ContextNotFound
+import uk.gov.hmrc.apigateway.model.ApiDefinition
+import uk.gov.hmrc.play.binding.PlayBindings.apiDefinitionFormat
+
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
+
+@Singleton
+class ApiDefinitionConnector @Inject()(wsClient: WSClient)
+  extends ServiceConnector(wsClient, "api-definition") {
+
+  def getByContext(context: String): Future[ApiDefinition] =
+    get[ApiDefinition](s"$serviceName?context=$context") recover {
+      case error: RuntimeException => throw ContextNotFound()
+    }
+
+}
