--- app/uk/gov/hmrc/connector/ServiceConnector.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.connector
-
-import play.api.libs.json.Format
-import play.api.libs.ws.WSClient
-import uk.gov.hmrc.play.config.ServicesConfig
-
-import scala.concurrent.Future
-
-abstract class ServiceConnector(wsClient: WSClient, val serviceName: String)
-  extends AbstractConnector(wsClient) with ServicesConfig {
-
-  private lazy val serviceBaseUrl = baseUrl(serviceName)
-
-  override def get[T](urlPath: String)(implicit format: Format[T]): Future[T] =
-    super.get(s"$serviceBaseUrl/$urlPath")
-
-}
