--- app/uk/gov/hmrc/controller/ProxyController.scala
@@ -1,43 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.controller
-
-import javax.inject.{Inject, Singleton}
-
-import play.Logger
-import play.api.libs.json.Json.toJson
-import play.api.mvc.Results._
-import play.api.mvc.{Action, BodyParsers}
-import uk.gov.hmrc.exception.GatewayError._
-import uk.gov.hmrc.play.binding.PlayBindings._
-import uk.gov.hmrc.service.ProxyService
-
-import scala.concurrent.ExecutionContext.Implicits.global
-
-@Singleton
-class ProxyController @Inject()(proxyService: ProxyService) {
-
-  def proxy = Action.async(BodyParsers.parse.anyContent) { implicit request =>
-    proxyService.proxy(request) recover {
-      case error => {
-        Logger.error("unexpected error", error)
-        InternalServerError(toJson(ServerError()))
-      }
-    }
-  }
-
-}
