--- app/uk/gov/hmrc/exception/GatewayError.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.exception
-
-class GatewayError(val code: String, val message: String) extends Throwable(message)
-
-object GatewayError {
-
-  case class ServerError() extends GatewayError("SERVER_ERROR", "Service unavailable")
-
-  case class ContextNotFound() extends GatewayError("NOT_FOUND", "Requested resource could not be found")
-
-  case class MatchingResourceNotFound() extends GatewayError("MATCHING_RESOURCE_NOT_FOUND", "A resource with the name in the request cannot be found in the API")
-
-  case class InvalidAcceptHeader() extends GatewayError("ACCEPT_HEADER_INVALID", "The accept header is missing or invalid")
-
-}
