--- app/uk/gov/hmrc/model/ApiDefinition.scala
@@ -1,25 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.model
-
-case class ApiDefinition(context: String, serviceBaseUrl: String, versions: Seq[ApiVersion])
-
-case class ApiVersion(version: String, endpoints: Seq[ApiEndpoint])
-
-case class ApiEndpoint(uriPattern: String, method: String, scope: Option[String] = None)
-
-case class ApiDefinitionMatch(context: String, serviceBaseUrl: String, apiVersion: String, scope: Option[String])
