--- app/uk/gov/hmrc/play/filter/PlayFilters.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.play.filter
-
-import javax.inject.{Inject, Singleton}
-
-import play.api.Environment
-import play.api.http.HttpFilters
-
-@Singleton
-class PlayFilters @Inject()(environment: Environment, proxyFilter: ProxyFilter) extends HttpFilters {
-
-  override val filters = Seq(proxyFilter)
-
-}
