--- app/uk/gov/hmrc/service/ProxyService.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.service
-
-import javax.inject.{Inject, Singleton}
-
-import play.api.mvc.{AnyContent, Request, Result}
-import uk.gov.hmrc.connector.impl.ProxyConnector
-import uk.gov.hmrc.util.HttpHeaders._
-
-import scala.concurrent.Future
-
-@Singleton
-class ProxyService @Inject()(proxyConnector: ProxyConnector) {
-
-  def proxy(request: Request[AnyContent]): Future[Result] =
-    proxyConnector.proxy(request, request.tags(X_API_GATEWAY_ENDPOINT))
-
-}
