+++ conf/application.conf
@@ -1,5 +1,5 @@
-play.http.requestHandler = uk.gov.hmrc.play.handler.ProxyRequestHandler
-play.http.filters = uk.gov.hmrc.play.filter.PlayFilters
+play.http.requestHandler = uk.gov.hmrc.apigateway.play.handler.ProxyRequestHandler
+play.http.filters = uk.gov.hmrc.apigateway.play.filter.PlayFilters
 
 microservice {
   services {
